requirements.txt

ipykernel
dspy
jupyter
python-dotenv
langchain
langchain-experimental
langchain-openai
pypdf
dspy-ai[faiss-cpu]
sentence-transformers
streamlit